fix: series page 위치 수정
